Update to newer libraries. Fix a bad unit test that had the wrong signature for Mockito exposed by the Mockito upgrade. Still need to upgrade Kafka libs (this is a bigger lift).

+++ connectors/kafka/src/test/java/forklift/message/KafkaMessageTests.java
@@ -31,21 +31,21 @@ public class KafkaMessageTests {
 
     @Test
     public void beforeProcessingFalseTest() throws ConnectorException, InterruptedException {
-        when(controller.acknowledge(eq(record), any(Integer.class))).thenReturn(false);
+        when(controller.acknowledge(eq(record), any(Long.class))).thenReturn(false);
 
         boolean acknowledged = message.beforeProcessing();
 
         assertFalse(acknowledged);
-        verify(controller).acknowledge(eq(record), any(Integer.class));
+        verify(controller).acknowledge(eq(record), any(Long.class));
     }
 
     @Test
     public void beforeProcessingCallsControllerSuccess() throws ConnectorException, InterruptedException {
-        when(controller.acknowledge(eq(record), any(Integer.class))).thenReturn(true);
+        when(controller.acknowledge(eq(record), any(Long.class))).thenReturn(true);
 
         boolean acknowledged = message.beforeProcessing();
 
         assertTrue(acknowledged);
-        verify(controller).acknowledge(eq(record), any(Integer.class));
+        verify(controller).acknowledge(eq(record), any(Long.class));
     }
 }
